Fix bugs with follow DAO

+++ server/src/main/java/com/cs204/server/dao/StoryDAO.java
@@ -6,7 +6,7 @@ import edu.byu.cs.tweeter.model.domain.Status;
 
 public interface StoryDAO {
     DataPage<Status> getPageOfStories(String alias, int pageSize, String lastPosted);
-    void setStory(String post, String user, Integer time, List<String> urls, List<String> mentions);
+    void setStory(String post, String user, Long time, List<String> urls, List<String> mentions);
     Status getStory(String alias);
     void deleteStory(String alias);
 }
